# src/main/spec/api.yml
openapi: 3.0.3
info:
  title: spring-api
  version: v1
servers:
  - url: https://localhost:8080
paths:
  /api/v1/brewery:
    post:
      summary: create brewery
      description: Method to create a new brewery.
      operationId: createBrewery
      tags:
        - Brewery
      parameters:
        - $ref: "#/components/parameters/apiKeyRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreweryNew"
      security:
        - ApiKeyAuth: [ ]
      responses:
        "201":
          description: "Response object"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brewery"
        default:
          description: "unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: get breweries
      description: Method to get a list of all breweries in the system.
      operationId: readBreweries
      tags:
        - Brewery
      parameters:
        - $ref: "#/components/parameters/apiKeyRequired"
      security:
        - ApiKeyAuth: [ ]
      responses:
        "200":
          description: "Response object"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreweryList"
        default:
          description: "unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    apiKeyRequired:
      in: "header"
      name: "X-API-Key"
      description: "uuid key of the tenant"
      required: true
      schema:
        type: "string"
        format: "uuid"
  schemas:
    BreweryNew:
      description: DTO for creating a new example.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 254
    Brewery:
      description: DTO with all fields.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 254
        lock_version:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BreweryList:
      description: Listing of examples.
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Brewery'
    ErrorArgument:
      description: Arguments used in the error message.
      properties:
        key:
          type: "string"
        value:
          type: "string"
        type:
          type: "string"
      required:
        - "key"
        - "type"
    Error:
      description: Custom error object with translated messages.
      properties:
        time:
          type: "string"
          format: "date-time"
        code:
          type: "string"
        title:
          type: "string"
        message:
          type: "string"
        args:
          type: "array"
          items:
            $ref: "#/components/schemas/ErrorArgument"
      required:
        - "time"
        - "code"
        - "title"
        - "message"